@@ -14,28 +14,29 @@ description: Quran Search Engine and Statistics.
 version: 0.0.0+0
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   # wisebay_essentials:
   #   path: c:/wisebay_essentials/
-  sqflite: any
-  rxdart: ^0.23.0
-  provider: ^4.0.5
-  charts_flutter: ^0.5.0
-  path_provider: any
-  share: ^0.6.4+3
-  launch_review: ^2.0.0
-  package_info: ^0.4.1
-  scrollable_positioned_list: ^0.1.8
-  arabic_numbers: ^0.0.1
+  sqflite: ^2.0.0+3
+  rxdart: ^0.27.1
+  provider: ^5.0.0
+  charts_flutter: ^0.11.0
+  path_provider: ^2.0.2
+  share: ^2.0.4
+  launch_review: ^3.0.1
+  package_info: ^2.0.2
+  scrollable_positioned_list: ^0.2.0-nullsafety.0
+  arabic_numbers: ^1.0.0
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  shared_preferences: ^0.5.7+3
-  http: ^0.12.2
+  cupertino_icons: ^1.0.3
+  shared_preferences: ^2.0.6
+  http: ^0.13.3
+  logging: ^1.0.1
 
 flutter_icons:
   image_path: 'icon/icon.png'
@@ -43,7 +44,7 @@ flutter_icons:
   ios: true
 
 dev_dependencies:
-  flutter_launcher_icons: ^0.7.4
+  flutter_launcher_icons: ^0.9.1
   flutter_test:
     sdk: flutter
 
