@@ -1,19 +1,41 @@
 import 'package:Yatadabaron/services/arabic-numbers-service.dart';
 import 'package:share/share.dart';
-
 import '../dtos/search-settings.dart';
 import '../dtos/verse-dto-collection.dart';
 import '../dtos/verse-dto.dart';
 import '../helpers/localization.dart';
 import '../helpers/utils.dart';
+import '../enums/enums.dart';
 
 class SearchSessionPayload {
-  final String chapterName;
+  final String? chapterName;
   final List<VerseDTO> results;
   final SearchSettings settings;
   SearchSessionPayload(this.settings, this.chapterName, this.results);
 
   String get summary {
+    int count = 0;
+    for (var result in results) {
+      List<String> words = result.verseText!.split(" ");
+      for (var word in words) {
+        bool found = false;
+        switch (settings.mode) {
+          case SearchMode.WORD:
+            found = word == this.settings.keyword;
+            break;
+          case SearchMode.START:
+          case SearchMode.END:
+          case SearchMode.WITHIN:
+            found = word.contains(this.settings.keyword);
+            break;
+          default:
+            break;
+        }
+        if (found) {
+          count++;
+        }
+      }
+    }
     if (settings.chapterID == 0) {
       String original = Localization.SEARCH_SUMMARY_WHOLE_QURAN;
       int chaptersCount = results.map((v) => v.chapterName).toSet().length;
@@ -21,9 +43,9 @@ class SearchSessionPayload {
         original,
         "#",
         [
-          ArabicNumbersService.insance.convert(results.length,reverse: false),
+          ArabicNumbersService.insance.convert(count, reverse: false),
           settings.keyword,
-          ArabicNumbersService.insance.convert(chaptersCount,reverse: false),
+          ArabicNumbersService.insance.convert(chaptersCount, reverse: false),
         ],
       );
       return original;
@@ -33,7 +55,7 @@ class SearchSessionPayload {
         original,
         "#",
         [
-          ArabicNumbersService.insance.convert(results.length,reverse: false),
+          ArabicNumbersService.insance.convert(count, reverse: false),
           settings.keyword,
           chapterName,
         ],
@@ -44,9 +66,9 @@ class SearchSessionPayload {
 
   List<VerseCollection> get verseCollections {
     List<VerseCollection> collections = [];
-    List<String> chapterNames =
+    List<String?> chapterNames =
         results.map((v) => v.chapterName).toSet().toList();
-    chapterNames.forEach((String chapter) {
+    chapterNames.forEach((String? chapter) {
       List<VerseDTO> verses =
           results.where((v) => v.chapterName == chapter).toList();
       collections.add(VerseCollection(verses, chapter));
@@ -55,16 +77,18 @@ class SearchSessionPayload {
     return collections.reversed.toList();
   }
 
-  void copyAll(){
+  void copyAll() {
     String resultsStr = "$summary\n\n";
     results.forEach((VerseDTO verseDTO) {
-      resultsStr += "${verseDTO.chapterName}\n${verseDTO.verseTextTashkel} {${verseDTO.verseID}}\n\n";
+      resultsStr +=
+          "${verseDTO.chapterName}\n${verseDTO.verseTextTashkel} {${verseDTO.verseID}}\n\n";
     });
     Share.share(resultsStr);
   }
 
-  void copyVerse(VerseDTO verseDTO){
-    String toCopy = "${verseDTO.chapterName}\n${verseDTO.verseTextTashkel} {${verseDTO.verseID}}";
+  void copyVerse(VerseDTO verseDTO) {
+    String toCopy =
+        "${verseDTO.chapterName}\n${verseDTO.verseTextTashkel} {${verseDTO.verseID}}";
     Share.share(toCopy);
   }
 }
