@@ -1,3 +1,5 @@
+import 'package:Yatadabaron/services/custom-prefs.dart';
+
 import '../blocs/generic-bloc.dart';
 import '../dtos/chapter-full-dto.dart';
 import '../dtos/chapter-simple-dto.dart';
@@ -6,8 +8,8 @@ import '../repositories/chapters-repository.dart';
 import '../repositories/verses-repository.dart';
 
 class MushafBloc {
-  MushafBloc(int chapterId, int verseId) {
-    selectChapter(chapterId, verseId);
+  MushafBloc(int? chapterId, int? verseId) {
+    reloadVerses(chapterId, verseId);
   }
 
   GenericBloc<List<VerseDTO>> _versesBloc = GenericBloc();
@@ -16,15 +18,33 @@ class MushafBloc {
   Stream<ChapterFullDTO> get selectedChapterStream =>
       _selectedChapterBloc.stream;
   Stream<List<VerseDTO>> get versesStream => _versesBloc.stream;
-  Future selectChapter(int chapterId, int verseId) async {
+  Future saveBookmark(int chapterId, int verseId) async {
+    await CustomSharedPreferences.instance.setBookmarkChapter(chapterId);
+    await CustomSharedPreferences.instance.setBookmarkVerse(verseId);
+    await reloadVerses(
+      chapterId,
+      verseId,
+    );
+  }
+
+  Future reloadVerses(int? chapterId, int? verseId) async {
     chapterId = chapterId ?? 1;
     ChapterFullDTO chapter =
         await ChaptersRepository.instance.getFullChapterById(chapterId);
     List<VerseDTO> verses =
         await VersesRepository.instance.getVersesByChapterId(chapterId, false);
+
     _selectedChapterBloc.add(chapter);
     if (verseId != null && verseId > 0) {
-      verses.firstWhere((v) => v.verseID == verseId)?.isSelected = true;
+      verses.firstWhere((v) => v.verseID == verseId).isSelected = true;
+    }
+    //Load the bookmarks
+    int? bmC = await CustomSharedPreferences.instance.getBookmarkChapter();
+    if (bmC != null) {
+      if (bmC == chapterId) {
+        int? bmV = await CustomSharedPreferences.instance.getBookmarkVerse();
+        verses.firstWhere((v) => v.verseID == bmV).isBookmark = true;
+      }
     }
     _versesBloc.add(verses);
   }
