@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 class Theming {
   static ThemeData darkTheme() {
-    Color accent = Color(0xfffff0cf);
+    Color accent = Color(0xfffbff00);
     return ThemeData.dark().copyWith(
         textTheme: ThemeData.dark().textTheme.apply(fontFamily: "Usmani"),
         accentColor: accent,
