@@ -4,13 +4,9 @@ class ArabicNumbersService {
   static ArabicNumbersService insance = ArabicNumbersService._();
   ArabicNumbersService._();
 
-  ArabicNumbers _converter;
+  ArabicNumbers _converter = ArabicNumbers();
 
-  void initialize() {
-    _converter = ArabicNumbers();
-  }
-
-  String convert(int x, {bool reverse = true}) {
+  String convert(int? x, {bool reverse = true}) {
     String converted = _converter.convert(x);
     if (reverse == true) {
       String reversed = converted.split('').reversed.join();
